- Changed package name course; - Added AbstractFilter
