--- app/src/main/java/by/itacademy/mikhalevich/icours/AverageSalary.java
@@ -1,51 +1,0 @@
-package by.itacademy.mikhalevich.icours;
-
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@WebServlet(value = "/average-salary")
-public class AverageSalary extends HttpServlet {
-
-    /*private List<Teacher> teachers;
-
-    @Override
-    public void init() throws ServletException {
-        super.init();
-        teachers = initModel();
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.setContentType("text/html");
-        resp.setCharacterEncoding("UTF-8");
-        PrintWriter writer = resp.getWriter();
-        writer.write("<p><span style='color: blue;'>Средняя зарплата следующих преподавателей: "
-                + averageSalary(teachers)+ "</span></p>");
-        teachers.forEach(teacher -> writer.write(teacher.getInfo() + "</p>"));
-    }
-
-    private BigDecimal averageSalary(List<Teacher> teachers) {
-        List<Integer> salaries = teachers.stream()
-                .map(Teacher::getSalary)
-                .collect(Collectors.toList());
-        return Accounting.average(salaries).setScale(2, RoundingMode.HALF_UP);
-    }
-
-    private List<Teacher> initModel() {
-        return List.of(
-                new Teacher("John", 30, 2000),
-                new Teacher("John", 35, 1100),
-                new Teacher("John", 30, 1200)
-        );
-    }*/
-}
