+++ app/src/main/java/by/itacademy/mikhalevich/icourse/servlet/page/AllStudentsController.java
@@ -1,0 +1,23 @@
+package by.itacademy.mikhalevich.icourse.servlet.page;
+
+import by.itacademy.mikhalevich.icourse.model.Student;
+import by.itacademy.mikhalevich.icourse.servlet.AbstractController;
+import by.itacademy.mikhalevich.icourse.util.RoutingUtils;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.List;
+
+@WebServlet("/students")
+public class AllStudentsController extends AbstractController {
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        List<Student> students = getStudentService().listAllStudents();
+        req.setAttribute("students", students);
+        RoutingUtils.forwardToPage("teachers.jsp", req, resp);
+    }
+}
