+++ app/src/main/java/by/itacademy/mikhalevich/icourse/servlet/page/ErrorController.java
@@ -1,0 +1,21 @@
+package by.itacademy.mikhalevich.icourse.servlet.page;
+
+import by.itacademy.mikhalevich.icourse.servlet.AbstractController;
+import by.itacademy.mikhalevich.icourse.util.RoutingUtils;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+
+@WebServlet("/error")
+public class ErrorController extends AbstractController {
+
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		RoutingUtils.forwardToPage("error.jsp", req, resp);
+	}
+}
