--- logic/src/main/java/by/itacademy/mikhalevich/icours/logic/StudentService.java
@@ -1,9 +1,0 @@
-package by.itacademy.mikhalevich.icours.logic;
-
-import by.itacademy.mikhalevich.icours.model.Student;
-
-import java.util.List;
-
-public interface StudentService {
-    List<Student> listAllStudents();
-}
