--- logic/src/main/java/by/itacademy/mikhalevich/icours/logic/exception/LogicalServerErrorException.java
@@ -1,17 +1,0 @@
-package by.itacademy.mikhalevich.icours.logic.exception;
-
-public class LogicalServerErrorException extends RuntimeException {
-
-	public LogicalServerErrorException(String message) {
-		super(message);
-	}
-
-	public LogicalServerErrorException(Throwable cause) {
-		super(cause);
-	}
-
-	public LogicalServerErrorException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
