+++ logic/src/main/java/by/itacademy/mikhalevich/icourse/logic/StudentService.java
@@ -1,0 +1,9 @@
+package by.itacademy.mikhalevich.icourse.logic;
+
+import by.itacademy.mikhalevich.icourse.model.Student;
+
+import java.util.List;
+
+public interface StudentService {
+    List<Student> listAllStudents();
+}
