+++ logic/src/main/java/by/itacademy/mikhalevich/icourse/logic/exception/LogicalServerErrorException.java
@@ -1,0 +1,17 @@
+package by.itacademy.mikhalevich.icourse.logic.exception;
+
+public class LogicalServerErrorException extends RuntimeException {
+
+	public LogicalServerErrorException(String message) {
+		super(message);
+	}
+
+	public LogicalServerErrorException(Throwable cause) {
+		super(cause);
+	}
+
+	public LogicalServerErrorException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
