--- model/src/main/java/by/itacademy/mikhalevich/icours/model/Person.java
@@ -1,65 +1,0 @@
-package by.itacademy.mikhalevich.icours.model;
-
-import java.util.Objects;
-
-public abstract class Person implements Printable {
-    private long id;
-    private String name;
-    private int age;
-
-    public Person() {
-    }
-
-    public Person(long id, String name, int age) {
-        this.id = id;
-        this.name = name;
-        this.age = age;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Person person = (Person) o;
-        return id == person.id &&
-                Objects.equals(name, person.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, age);
-    }
-
-    @Override
-    public String toString() {
-        return "Person{" +
-                "id='" + id + '\'' +
-                "name='" + name + '\'' +
-                ", age=" + age +
-                '}';
-    }
-}
