--- model/src/main/java/by/itacademy/mikhalevich/icours/model/Printable.java
@@ -1,5 +1,0 @@
-package by.itacademy.mikhalevich.icours.model;
-
-public interface Printable {
-    public String getInfo();
-}
