+++ model/src/main/java/by/itacademy/mikhalevich/icourse/model/Printable.java
@@ -1,0 +1,5 @@
+package by.itacademy.mikhalevich.icourse.model;
+
+public interface Printable {
+    public String getInfo();
+}
